backend/pyproject.toml
[tool.poetry]
name = "glavcontrol"
version = "0.1.0"
description = ""
authors = ["Nikita Orekhov <nikanor97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.10.4"
sqlalchemy = "1.4.41"  # Version restricted by sqlmodel
python-dotenv = "^1.0.0"
fastapi = "^0.95.1"
pydantic = "^1.10.7"
psycopg2-binary = "^2.9.6"
sqlmodel = "^0.0.8"
loguru = "^0.7.0"
uvicorn = "^0.22.0"
uvloop = "^0.17.0"
asyncpg = "^0.27.0"
aiofiles = "^23.1.0"
python-multipart = "^0.0.6"
orjson = "^3.8.12"
#ffmpeg-python = "^0.2.0"
types-aiofiles = "^23.1.0.2"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.0.1"
types-passlib = "^1.7.7.12"
types-python-jose = "^3.3.4.7"
jinja2 = "^3.1.2"
sqlalchemy-utils = "^0.41.1"
pandas = "^2.0.1"
aiohttp = "^3.8.4"
aio-pika = "^9.0.7"
ujson = "^5.7.0"
#opencv-python = "^4.7.0.72"
#torch = "^2.0.1"
#plotly = "^5.14.1"
openpyxl = "^3.1.2"
faker = "^25.8.0"
xlsxwriter = "^3.2.0"
greenlet = "^3.0.3"

[tool.poetry.dev-dependencies]
pre-commit = "^3.3.1"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"
pytest = "^7.3.1"
pycln = "^2.1.3"
notebook = "^6.5.4"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
